Update README.md
History

Ham Radio Deluxe is software package that has multiple uses for the radio amateur. It was originally developed by Simon Brown (HB9DRV) around 2003 and was distributed as freeware. The software can interface with most transceivers that would allow control from the software. It allows the amateur radio operator to log contacts into a database that could be used later to verify contacts through various online and offline award systems. The software also includes a "Digital" software that allows encoding and decoding of various AFSK, FSK, CW and other digital messages. This is just a fraction of what the software had to offer. The software would combine multiple actions to allow for smoother and more enjoyable on-air experience with HRD doing most of the work.

In 2011, Simon's attention went from HRD to another project (SDR-Radio) and it was possible that development would cease. At this point, three amateur radio operators got together and purchased the source code from Simon. Since many people had financially supported HRD through donations, the last version (5.X) had to remain free.

In 2013, The new owners and developers at HRD released version 6.0. However it's no longer free. HRDsoft was now charging $99 for the software and 1 year of support. In order to obtain updates and support after the first year, the user had to pay $50 annually. HRD has since made improvements to the software and sold many units. Most of the customers were satisfied but there were bugs and issues with both software and hardware. There were customers who had issues with the software and allegedly received poor support from the staff.